ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axiom-mobile-ml"
version = "0.1.0"
description = "Baseline ML utilities and experiment runners for AXIOM-Mobile."
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
train = [
    "torch>=2.1",
    "Pillow>=10.0",
    "numpy>=1.24",
]
dev = []

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
